feature/colormanager: rename color_pairs to color_manager
