+++ suplemon/color_manager.py
